v2.5

+++ omsa_install.sh
@@ -232,14 +232,17 @@ sleep 1
 
 if [ $DEBUG -eq 1 ]
 then
-  echo -e "\e[96m++ $PHASE - rm /etc/apt/sources.list.d/linux.dell.com.sources.list"
+  echo -e "\e[96m++ $PHASE - [[ -f /etc/apt/sources.list.d/linux.dell.com.sources.list ]] && rm /etc/apt/sources.list.d/linux.dell.com.sources.list"
   echo -e "\e[96m++ $PHASE - mkdir /opt/dell/srvadmin/sbin\e[39m"
   echo -e "\e[96m++ $PHASE - apt purge srvadmin-*\e[39m"
 else
   echo
-  rm /etc/apt/sources.list.d/linux.dell.com.sources.list
-  mkdir /opt/dell/srvadmin/sbin
-  apt purge srvadmin-*
+  [[ -f "/etc/apt/sources.list.d/linux.dell.com.sources.list" ]] && rm /etc/apt/sources.list.d/linux.dell.com.sources.list
+  [[ ! -d "/opt/dell/srvadmin/sbin" ]] && mkdir /opt/dell/srvadmin/sbin
+
+  if dpkg --get-selections | grep -q "^srvadmin-*[[:space:]]*install$" >/dev/null; then
+    apt purge srvadmin-*
+  fi
 fi
 
 ### End Phase 1
